crawl-netatmo/requirements.txt
oauthlib
google-cloud-bigquery[pandas,pyarrow]